Made a connection to the database. Pulled up one servlet.

+++ src/main/java/repository/ProductRepository.java
@@ -2,6 +2,6 @@ package repository;
 
 import model.Product;
 
-public interface ProductRepository extends Repository<Product, Long> {
+public interface ProductRepository extends Repository<Product, Integer> {
 
 }
