+++ src/main/java/repository/ProductWarehouseRepository.java
@@ -5,6 +5,6 @@ import model.ProductWarehouse;
 
 import java.util.Optional;
 
-public interface ProductWarehouseRepository extends Repository<ProductWarehouse, Long>{
+public interface ProductWarehouseRepository extends Repository<ProductWarehouse, Integer>{
     Optional<ProductWarehouse> getByProduct(Product product);
 }
