+++ src/main/java/repository/ReceiptRepository.java
@@ -2,5 +2,5 @@ package repository;
 
 import model.Receipt;
 
-public interface ReceiptRepository  extends Repository<Receipt, Long> {
+public interface ReceiptRepository  extends Repository<Receipt, Integer> {
 }
